requirements.txt
# Python dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1

# Data processing
pandas==2.1.4
numpy==1.24.3

# Web scraping and API clients
httpx==0.25.2
beautifulsoup4==4.12.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1